requirements/dev.txt
flask
flask-sqlalchemy
psycopg2
Pillow
numpy
flask-jwt-extended